+++ backend/garpix_user/admin/__init__.py
@@ -4,5 +4,11 @@ from .user_session import UserSessionAdmin  # noqa
 
 from .user import UserAdmin  # noqa
 
+from .password_history import PasswordHistoryAdmin  # noqa
+
+
 if settings.GARPIX_USER.get('USE_REFERRAL_LINKS', False):
     from .referral_type import ReferralTypeAdmin  # noqa
+
+if settings.GARPIX_USER.get('ADMIN_PASSWORD_SETTINGS', False):
+    from .site_config import GarpixUserPasswordConfigurationAdmin  # noqa