+++ backend/garpix_user/admin/password_history.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+
+from ..models import PasswordHistory
+
+
+@admin.register(PasswordHistory)
+class PasswordHistoryAdmin(admin.ModelAdmin):
+    list_display = ['user', 'created_at']
+
+    def has_delete_permission(self, request, obj=None):
+        return False
+
+    def has_change_permission(self, request, obj=None):
+        return False