+++ backend/garpix_user/models/__init__.py
@@ -3,3 +3,5 @@ from .access_token import AccessToken  # noqa
 from .user_session import UserSession # noqa
 from .refferal import ReferralType, ReferralUserLink  # noqa
 from .user import GarpixUser  # noqa
+from .site_config import GarpixUserPasswordConfiguration  # noqa
+from .password_history import PasswordHistory  # noqa