+++ backend/user/models/user.py
@@ -3,7 +3,7 @@ from garpix_user.models import GarpixUser
 
 class User(GarpixUser):
 
-    USERNAME_FIELDS = ('phone', 'email')
+    USERNAME_FIELDS = ('phone', 'email', 'username')
 
     class Meta:
         verbose_name = 'Пользователь'